move some widgets to lib/widgets folder
