+++ lib/widgets/animated_content_builder.dart
